+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Do not redirect mercurial hook requests
+
 ## 2.0.0 - 2020-06-04
 ### Changed
 - Changeover to MIT license ([#2](https://github.com/scm-manager/scm-cas-plugin/pull/2))